Ajustes generales

+++ api/src/routes/app/cost/quotes/routePaymentMethods.php
@@ -1,8 +1,10 @@
 <?php
 
+use tezlikv3\dao\GeneralQuotesDao;
 use tezlikv3\dao\PaymentMethodsDao;
 
 $paymentMethodsDao = new PaymentMethodsDao();
+$generalQuotesDao = new GeneralQuotesDao();
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
@@ -52,8 +54,17 @@ $app->post('/updatePaymentMethod', function (Request $request, Response $respons
     return $response->withStatus(200)->withHeader('Content-Type', 'application/json');
 });
 
-$app->get('/deletePaymentMethod/{id_method}', function (Request $request, Response $response, $args) use ($paymentMethodsDao) {
-    $paymentMethods = $paymentMethodsDao->deletePaymentMethod($args['id_method']);
+$app->get('/deletePaymentMethod/{id_method}', function (Request $request, Response $response, $args) use (
+    $paymentMethodsDao,
+    $generalQuotesDao
+) {
+
+    $quotes = $generalQuotesDao->findPaymentMethod($args['id_method']);
+
+    if (sizeof($quotes) > 0)
+        $paymentMethods = $paymentMethodsDao->changeFlagPaymentMethod($args['id_method'], 1);
+    else
+        $paymentMethods = $paymentMethodsDao->deletePaymentMethod($args['id_method']);
 
     if ($paymentMethods == null)
         $resp = array('success' => true, 'message' => 'Metodo de pago eliminado correctamente');
